Fix errors

+++ Recipe.cs
@@ -7,10 +7,10 @@ using System.Threading.Tasks;
 namespace LemonadeStand
 {
     public class Recipe : Inventory
-    {
-        
-        public int pitchers;
-        public int CreateRecipe()
+    {        
+        public double pitchers;
+
+        public double CreateRecipe()
         {
             Console.WriteLine("Now that you have your ingredients, it's time to make your lemonade. \nHow many lemons will you use in today's batch?");
             lemonsInRecipe = double.Parse(Console.ReadLine());
@@ -19,7 +19,7 @@ namespace LemonadeStand
             Console.WriteLine("How many ice cubes will you use?");
             iceInRecipe = double.Parse(Console.ReadLine());
             Console.WriteLine($"Your recipe today is {lemonsInRecipe} lemons, {sugarInRecipe} cups of sugar, and {iceInRecipe} ice cubes. \nHow many pitchers will you make using these ingredients? Keep in mind each pitcher can hold 8 cups of lemonade. \nPress enter to continue.");
-            pitchers = int.Parse(Console.ReadLine());
+            pitchers = double.Parse(Console.ReadLine());
             //setPrice.CalcCostPerCup(pitchers); //TODO
             //TrackIngredients();
             return pitchers;
